ai_agent/requirements.txt
# AI Agent for Feed Products Data
# Core dependencies

# Google Gemini AI SDK
google-generativeai>=0.8.0

# Language detection and translation
langdetect>=1.0.9
deep-translator>=1.11.4

# Database
sqlalchemy>=2.0.0

# Web API (for future UI integration)
fastapi>=0.115.0
uvicorn>=0.32.0

# CLI and formatting
rich>=13.9.0
python-dotenv>=1.0.0

# Data processing
pandas>=2.2.0

# Async support
aiohttp>=3.10.0
